fix(assertions): co-locate character count handler
